08052025 add exceptions delete users

+++ src/main/java/com/andre/usandoSpring/services/UserService.java
@@ -2,9 +2,12 @@ package com.andre.usandoSpring.services;
 
 import com.andre.usandoSpring.entities.User;
 import com.andre.usandoSpring.repositories.UserRepository;
+import com.andre.usandoSpring.services.exceptions.DatabaseException;
 import com.andre.usandoSpring.services.exceptions.ResourceNotFoundException;
 import org.hibernate.annotations.Comment;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
@@ -17,24 +20,34 @@ public class UserService {
     @Autowired
     private UserRepository repository;
 
-    public List<User> findAll(){
-      return repository.findAll();
-    };
+    public List<User> findAll() {
+        return repository.findAll();
+    }
 
-    public User findById(Long id){
+    ;
+
+    public User findById(Long id) {
         Optional<User> obj = repository.findById(id);
         return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
-    public User insert(User user){
+    public User insert(User user) {
         return repository.save(user);
     }
 
-    public void delete(Long id){
-       repository.deleteById(id);
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
-    public User update(Long id, User user){
+    public User update(Long id, User user) {
         User entity = repository.getReferenceById(id);
         updateData(entity, user);
         return repository.save(entity);
